bitcoin: move BTC_DECIMAL to const
