+++ packages/xchain-bitcoin/src/const.ts
@@ -4,3 +4,5 @@
  * @see https://github.com/bitcoin/bitcoin/blob/db88db47278d2e7208c50d16ab10cb355067d071/src/validation.h#L56
  */
 export const MIN_TX_FEE = 1000
+
+export const BTC_DECIMAL = 8
