+++ packages/xchain-bitcoin/src/index.ts
@@ -1,11 +1,11 @@
 export * from './types'
 export * from './client'
+export * from './const'
 export {
   broadcastTx,
   getDefaultFees,
   getDefaultFeesWithRates,
   getPrefix,
-  BTC_DECIMAL,
   scanUTXOs,
   buildTx,
   validateAddress,
